# docs/swagger.yaml
basePath: /
definitions:
  article.DeleteComment:
    properties:
      article_id:
        type: integer
      comment_id:
        type: integer
    required:
    - article_id
    - comment_id
    type: object
  article.EditComment:
    properties:
      article_id:
        type: integer
      comment:
        minLength: 3
        type: string
      comment_id:
        type: integer
    required:
    - article_id
    - comment
    - comment_id
    type: object
  article.NewComment:
    properties:
      article_id:
        type: integer
      comment:
        minLength: 3
        type: string
    required:
    - article_id
    - comment
    type: object
...
  types.LikeOrDislikeArticle:
    properties:
      article_id:
        type: integer
    required:
    - article_id
    type: object
...
  user.EditUser:
    properties:
      first_name:
        type: string
      last_name:
        type: string
    type: object
info:
  contact: {}
  title: Swagger Example API
  version: "1.0"
paths:
  /article/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
      security:
      - BearerAuth: []
      summary: Delete Article By Id
      tags:
      - articles
    get:
      consumes:
      - application/json
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
      summary: Get Article By Id
      tags:
      - articles
  /article/add:
    post:
      consumes:
      - multipart/form-data
      description: Upload an article along with an image
      parameters:
      - description: Article Title
        in: formData
        name: title
        required: true
        type: string
      - description: Article Description
        in: formData
        name: description
        required: true
        type: string
      - description: Article Summary
        in: formData
        name: summary
        required: true
        type: string
      - description: Article Category
        in: formData
        name: category
        required: true
        type: string
      - description: Article Image
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      tags:
      - articles
...
  /article/byCategory/{categoryId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Category ID
        in: path
        name: categoryId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
      summary: Get Articles By category
      tags:
      - articles
  /article/comment/add:
    post:
      consumes:
      - application/json
      description: Add comment by user
      parameters:
      - description: Add new comment
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/article.NewComment'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      tags:
      - articles
  /article/comment/delete:
    delete:
      consumes:
      - application/json
      description: delete comment by user
      parameters:
      - description: Edit comment
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/article.DeleteComment'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      tags:
      - articles
  /article/comment/edit:
    put:
      consumes:
      - application/json
      description: edit comment by user
      parameters:
      - description: Edit comment
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/article.EditComment'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      tags:
      - articles
  /article/dislike:
    put:
      consumes:
      - application/json
      description: dislike article
      parameters:
      - description: like article
        in: body
        name: register
        schema:
          $ref: '#/definitions/types.LikeOrDislikeArticle'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      tags:
      - articles
  /article/edit:
    put:
      consumes:
      - multipart/form-data
      description: Edit an article
      parameters:
      - description: Article Title
        in: formData
        name: title
        type: string
      - description: Article Description
        in: formData
        name: description
        type: string
      - description: Article Summary
        in: formData
        name: summary
        type: string
      - description: Article Category
        in: formData
        name: category
        type: string
      - description: Article Status
        in: formData
        name: status
        type: string
      - description: Article id
        in: formData
        name: id
        type: string
      - description: Article Image
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      tags:
      - articles
  /article/like:
    put:
      consumes:
      - application/json
      description: like article
      parameters:
      - description: like article
        in: body
        name: register
        schema:
          $ref: '#/definitions/types.LikeOrDislikeArticle'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      tags:
      - articles
...
  /article/property/{propertyId}:
    delete:
      parameters:
      - description: Property ID
        in: path
        name: propertyId
        required: true
        type: integer
      responses: {}
      security:
      - BearerAuth: []
      tags:
      - articles
  /article/property/add:
    post:
      parameters:
      - description: Property description
        in: formData
        name: description
        type: string
      - description: Article id
        in: formData
        name: article_id
        required: true
        type: integer
      - description: Property image
        in: formData
        name: image
        type: file
      responses: {}
      security:
      - BearerAuth: []
      tags:
      - articles
  /article/property/edit:
    put:
      parameters:
      - description: Property description
        in: formData
        name: description
        type: string
      - description: Article id
        in: formData
        name: property_id
        required: true
        type: integer
      - description: Property image
        in: formData
        name: image
        type: file
      responses: {}
      security:
      - BearerAuth: []
      tags:
      - articles
...
  /user/edit:
    put:
      consumes:
      - application/json
      description: edit user data
      parameters:
      - description: Edit data
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/user.EditUser'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"